tecteonic: rename the C++ files with a "xetex-" prefix
Final file names are along the lines of "xetex-XeTeXFontMgr.h" Note that the
classes that these files define have names such as `XeTeXFontMgr`, so it seems
best to keep the "XeTeX" in the filename.
